pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "telescope"
version = "0.1.0"
description = "Telegram rule-based engine for matching and notifications"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "telethon==1.34.0",
  "python-dotenv==1.0.1",
  "art",
]

[project.scripts]
telescope = "app:main"

[tool.setuptools]
package-dir = {"" = "src"}
py-modules = ["app", "settings", "client", "get_session", "init"]

[tool.setuptools.packages.find]
where = ["src"]
